dom/indexedDB/test/test_file_cross_database_copying.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript;version=1.7">
  function testSteps()
  {
    const READ_WRITE = IDBTransaction.READ_WRITE;

    const databaseInfo = [
      { name: window.location.pathname + "1" },
      { name: window.location.pathname + "2" }
    ];

    const objectStoreName = "Blobs";

    const fileData = { key: 1, file: getRandomFile("random.bin", 100000) };

    let databases = [];
    for each (let info in databaseInfo) {
      let request = indexedDB.open(info.name, 1);
      request.onerror = errorHandler;
      request.onupgradeneeded = grabEventAndContinueHandler;
      request.onsuccess = grabEventAndContinueHandler;
      let event = yield;

      is(event.type, "upgradeneeded", "Got correct event type");

      let db = event.target.result;
      db.onerror = errorHandler;

      let objectStore = db.createObjectStore(objectStoreName, { });
      objectStore.add(fileData.file, fileData.key);

      event = yield;

      is(event.type, "success", "Got correct event type");

      databases.push(db);
    }

    let refResult;
    for each (let db in databases) {
      let request = db.transaction([objectStoreName])
                      .objectStore(objectStoreName).get(fileData.key);
      request.onsuccess = grabEventAndContinueHandler;
      event = yield;

      let result = event.target.result;
      verifyBlob(result, fileData.file, 1);
      yield;

      if (!refResult) {
        refResult = result;
        continue;
      }

      isnot(SpecialPowers.getMozFullPath(result),
            SpecialPowers.getMozFullPath(refResult), "Different os files");
    }

    for (let i = 1; i < databases.length; i++) {
      let db = databases[i];

      let objectStore = db.transaction([objectStoreName], READ_WRITE)
                          .objectStore(objectStoreName);

      request = objectStore.add(refResult, 2);
      request.onsuccess = grabEventAndContinueHandler;
      event = yield;

      is(event.target.result, 2, "Got correct key");

      request = objectStore.get(2);
      request.onsuccess = grabEventAndContinueHandler;
      event = yield;

      let result = event.target.result;
      verifyBlob(result, refResult, 2);
      yield;

      isnot(SpecialPowers.getMozFullPath(result),
            SpecialPowers.getMozFullPath(refResult), "Different os files");
    }

    is(bufferCache.length, 2, "Correct length");

    finishTest();
    yield;
  }
  </script>
  <script type="text/javascript;version=1.7" src="file.js"></script>
  <script type="text/javascript;version=1.7" src="helpers.js"></script>

</head>

<body onload="runTest();"></body>

</html>
